refactor: 177 - refactor quantity code
